+++ MAT_ZIP/MAT_ZIP/pom.xml
@@ -199,7 +199,7 @@
 		    <version>4.5.13</version>
 		</dependency>
 		<dependency>
-	     <groupId>org.json</groupId
+	     <groupId>org.json</groupId>
        <artifactId>json</artifactId>
 	     <version>20210307</version>
 		</dependency>
