+++ MAT_ZIP/MAT_ZIP/src/main/java/com/mat/zip/point/controller/PointExchangeHistoryController.java
@@ -1,8 +1,10 @@
-package com.mat.zip.point;
+package com.mat.zip.point.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
+
+import com.mat.zip.point.service.PointExchangeHistoryServiceImpl;
 
 @Controller
 public class PointExchangeHistoryController {
@@ -10,7 +12,7 @@ public class PointExchangeHistoryController {
 	@Autowired
 	PointExchangeHistoryServiceImpl PointExchangeHistoryService;
 
-	@RequestMapping("/exchange")
+	@RequestMapping("/point/exchange")
 	public void exchange(String user_id,  int id) {
 
 		PointExchangeHistoryService.exChange(user_id, id);
