+++ MAT_ZIP/MAT_ZIP/src/main/java/com/mat/zip/point/controller/gifticonController.java
@@ -1,4 +1,4 @@
-package com.mat.zip.point;
+package com.mat.zip.point.controller;
 
 import java.util.List;
 
@@ -7,13 +7,16 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import com.mat.zip.point.dao.gifticonDAO;
+import com.mat.zip.point.model.gifticonVO;
+
 @Controller
 public class gifticonController {
 
 	@Autowired
 	gifticonDAO dao;
 
-	@RequestMapping("gifticon")
+	@RequestMapping("/point/gifticon")
 	public void gifticon(Model model) {
 		List<gifticonVO> list = dao.findall();
 		model.addAttribute("list", list);
