+++ MAT_ZIP/MAT_ZIP/src/main/java/com/mat/zip/point/dao/gifticonDAO.java
@@ -1,10 +1,12 @@
-package com.mat.zip.point;
+package com.mat.zip.point.dao;
 
 import java.util.List;
 
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+
+import com.mat.zip.point.model.gifticonVO;
 
 @Component
 public class gifticonDAO {
