+++ MAT_ZIP/MAT_ZIP/src/main/java/com/mat/zip/point/model/SendSmsVO.java
@@ -1,4 +1,4 @@
-package com.mat.zip.point;
+package com.mat.zip.point.model;
 
 public class SendSmsVO {
