+++ MAT_ZIP/MAT_ZIP/src/main/java/com/mat/zip/point/service/PointExchangeHistoryService.java
@@ -1,4 +1,4 @@
-package com.mat.zip.point;
+package com.mat.zip.point.service;
 
 public interface PointExchangeHistoryService {
     void exChange(String userId, int id);
