+++ MAT_ZIP/MAT_ZIP/src/main/java/com/mat/zip/point/service/PointSaveHistoryService.java
@@ -1,4 +1,6 @@
-package com.mat.zip.point;
+package com.mat.zip.point.service;
+
+import com.mat.zip.point.model.PointSaveHistoryVO;
 
 public interface PointSaveHistoryService {
 	void addPoint(PointSaveHistoryVO bag2);
