+++ MAT_ZIP/MAT_ZIP/src/main/java/com/mat/zip/point/service/PointSaveHistoryServiceImpl.java
@@ -1,7 +1,10 @@
-package com.mat.zip.point;
+package com.mat.zip.point.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.mat.zip.point.dao.PointSaveHistoryDAO;
+import com.mat.zip.point.model.PointSaveHistoryVO;
 
 @Service
 public class PointSaveHistoryServiceImpl implements PointSaveHistoryService{
