+++ MAT_ZIP/MAT_ZIP/src/main/resources/mybatis-config.xml
@@ -17,11 +17,11 @@
       <typeAlias type="com.mat.zip.boss.return_CustomerCountVO" alias="return_CustomerCountVO"></typeAlias>
       <typeAlias type="com.mat.zip.boss.return_OrderCountVO" alias="return_OrderCountVO"></typeAlias>
       <typeAlias type="com.mat.zip.boss.return_OrderTotalVO" alias="return_OrderTotalVO"></typeAlias>
-		<typeAlias type="com.mat.zip.point.gifticonVO" alias="gifticonVO"></typeAlias>
-		<typeAlias type="com.mat.zip.point.ProductPointVO" alias="productPointVO"></typeAlias>
-		<typeAlias type="com.mat.zip.point.PointExchangeHistoryVO" alias="PointExchangeHistoryVO"></typeAlias>
-		<typeAlias type="com.mat.zip.point.PointSaveHistoryVO" alias="PointSaveHistoryVO"></typeAlias>
-		<typeAlias type="com.mat.zip.point.PointDetailHistoryVO" alias="PointDetailHistoryVO"></typeAlias>
+		<typeAlias type="com.mat.zip.point.model.gifticonVO" alias="gifticonVO"></typeAlias>
+		<typeAlias type="com.mat.zip.point.model.ProductPointVO" alias="productPointVO"></typeAlias>
+		<typeAlias type="com.mat.zip.point.model.PointExchangeHistoryVO" alias="PointExchangeHistoryVO"></typeAlias>
+		<typeAlias type="com.mat.zip.point.model.PointSaveHistoryVO" alias="PointSaveHistoryVO"></typeAlias>
+		<typeAlias type="com.mat.zip.point.model.PointDetailHistoryVO" alias="PointDetailHistoryVO"></typeAlias>
       
    </typeAliases>
