+++ MAT_ZIP/MAT_ZIP/src/main/webapp/point/gifticonTest.jsp
@@ -1,0 +1,12 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+	pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+</head>
+<body>
+<a href = "gifticon">기프티콘 목록 가져오기</a>
+</body>
+</html>
